requirements.txt
tikapi
fastapi
google-genai
google-cloud-bigquery
python-dotenv
tqdm
